requirements.txt
discord.py
dnspython
google-api-python-client
html2text
lxml
motor
numpy
opencv-python-headless
Pillow
psutil
pygsheets
PyOpenSSL
scipy
scikit-learn
uvloop
